fix task3.3.3.1
